+++ src/textual_fspicker/file_open.py
@@ -25,6 +25,7 @@ class FileOpen(BaseFileDialog):
         *,
         filters: Filters | None = None,
         must_exist: bool = True,
+        default_file: str | Path | None = None,
     ) -> None:
         """Initialise the `FileOpen` dialog.
 
@@ -33,8 +34,15 @@ class FileOpen(BaseFileDialog):
             title: Optional title.
             filters: Optional filters to show in the dialog.
             must_exist: Flag to say if the file must exist.
+            default_file: The default filename to place in the input.
         """
-        super().__init__(location, title, select_button="Open", filters=filters)
+        super().__init__(
+            location,
+            title,
+            select_button="Open",
+            filters=filters,
+            default_file=default_file,
+        )
         self._must_exist = must_exist
         """Must the file exist?"""