+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `default_file` to `FileOpen` and `FileSave`.
+
 ## v0.1.1
 
 **Released: 2025-01-16**