+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `default_file` to `FileOpen` and `FileSave`.
+  ([#18](https://github.com/davep/textual-fspicker/pull/18))
+
 ## v0.1.1
 
 **Released: 2025-01-16**